# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: FactorSage API
  version: "1.0.0"
  summary: Stock research, strategy design, backtesting and monitoring.
  description: |
    The HTTP API behind FactorSage.

    **Authentication** is a session cookie issued by `POST /auth/login` or by the Google OAuth
    callback. It is `HttpOnly`, so browser clients send it automatically and never read it; a
    non-browser client must store and resend it. There is no bearer token and no API key.

    **Authorization** has two independent axes, both resolved server-side from persisted state and
    never from anything the client sends: a role (`USER` / `ADMIN`) and a commercial plan
    (`FREE` / `STARTER` / `PRO`). A plan limit is refused with `403` and a stable
    `ENTITLEMENT_*` code. Ownership is not a separate axis — every user-owned collection is scoped
    to the caller, and somebody else's resource answers exactly like one that does not exist.

    **Rate limiting** is applied per endpoint class from a central policy catalog, keyed by user id
    for authenticated traffic and by client IP otherwise. Exceeding a policy answers `429` with
    `Retry-After`. See `info.x-rate-limit-policies`.

    **Errors** share one machine-readable shape: an HTTP `statusCode`, a human `message`, and — for
    every refusal a client may reasonably branch on — a stable `code`. Clients should branch on
    `code`, never on `message`.
  contact:
    name: FactorSage engineering
  license:
    name: UNLICENSED

...

...
servers:
...
paths:
...
  /strategies/{strategyId}:
    parameters:
      - { $ref: "#/components/parameters/StrategyId" }
    get:
      tags: [Strategies]
      operationId: getStrategy
      summary: One strategy with its full definition.
      security: [{ cookieAuth: [] }]
      x-rate-limit: { policy: standard-read }
      responses:
        "200":
          description: The strategy.
          headers: *rateLimitHeaders
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StrategyDetail" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/OwnedResourceNotFound" }
        "429": { $ref: "#/components/responses/TooManyRequests" }

    patch:
      tags: [Strategies]
      operationId: updateStrategy
      summary: Rename a strategy or change its description.
      description: Metadata only. The definition is replaced through its own endpoint.
      security: [{ cookieAuth: [] }]
      x-rate-limit: { policy: mutation }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateStrategyRequest" }
      responses:
        "200":
          description: The updated summary.
          headers: *rateLimitHeaders
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StrategySummary" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/OwnedResourceNotFound" }
        "429": { $ref: "#/components/responses/TooManyRequests" }

    delete:
      tags: [Strategies]
      operationId: deleteStrategy
      summary: Delete a strategy.
      description: |
        Completed backtest runs are unaffected: a run holds an immutable snapshot of the strategy it
        executed, so deleting the live strategy never deletes or reinterprets a result.
      security: [{ cookieAuth: [] }]
      x-rate-limit: { policy: mutation }
      responses:
        "204":
          description: Deleted.
          headers: *rateLimitHeaders
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/OwnedResourceNotFound" }
        "409":
          description: A monitor still references this strategy. Delete or rebind the monitor first.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
...
  /backtests:
    get:
      tags: [Backtests]
      operationId: listBacktestRuns
      summary: The caller's backtest runs.
      security: [{ cookieAuth: [] }]
      x-rate-limit: { policy: standard-read }
      responses:
        "200":
          description: Run summaries, newest first.
          headers: *rateLimitHeaders
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/BacktestRunSummary" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/TooManyRequests" }

    post:
      tags: [Backtests]
      operationId: submitBacktestRun
      summary: Submit a backtest for asynchronous execution.
      description: |
        Answers `202` with a queued run rather than blocking: a backtest is durable work a separate
        worker process executes, sometimes for minutes. Poll `/backtests/{runId}/progress`.

        Submission takes an immutable snapshot of the strategy, the list, the securities and their
        buy windows, the benchmark and the engine methodology. Nothing in a submitted run is ever
        re-derived from current rows, so editing or deleting any of them afterwards cannot change or
        invalidate the result.

        This is the most expensive thing a client can ask for, so it carries the tightest capacity
        policy in the catalog — twenty an hour per user, and sixty an hour per client address, the
        second bucket being what sees one origin spreading submissions across several accounts.
        That is separate from `backtests.maxConcurrentRuns`, which is a plan entitlement and bounds
        how many may execute at once.
      security: [{ cookieAuth: [] }]
      x-rate-limit: { policy: backtest-execution }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateBacktestRunRequest" }
            examples:
              default:
                value:
                  strategyId: "1f2e3d4c-5b6a-4790-8123-456789abcdef"
                  stockListId: "9a8b7c6d-5e4f-4321-8fed-cba987654321"
                  benchmarkCode: "SP500"
                  startDate: "2015-01-01"
                  endDate: "2024-12-31"
                  initialCapital: 100000
                  monthlyContribution: 1000
                  maximumPositions: 10
      responses:
        "202":
          description: The run was queued. Its status is `QUEUED` and no result exists yet.
          headers: *rateLimitHeaders
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BacktestRunDetail" }
        "400":
          description: |
            The configuration cannot be executed — an unknown or unowned strategy or list, an empty
            list, an out-of-range period, or a capital/allocation value outside its bounds. Carries
            `code: BACKTEST_INVALID`.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              examples:
                invalid:
                  value:
                    statusCode: 400
                    message: "The selected stock list has no securities."
                    code: BACKTEST_INVALID
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/EntitlementDenied" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "503":
          description: |
            The run cannot be created under the methodology it would have to record — for example
            the execution calendar is unavailable. Not the caller's fault, and the same request will
            work later.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /backtests/{runId}:
    parameters:
      - { $ref: "#/components/parameters/RunId" }
    get:
      tags: [Backtests]
      operationId: getBacktestRun
      summary: One run — its immutable configuration snapshot, progress and result.
      security: [{ cookieAuth: [] }]
      x-rate-limit: { policy: standard-read }
      responses:
        "200":
          description: |
            The run. `result` is populated only for `COMPLETED`; `failure` only for `FAILED`;
            `live` only while a simulation is in flight.
          headers: *rateLimitHeaders
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BacktestRunDetail" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/OwnedResourceNotFound" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /backtests/{runId}/progress:
    parameters:
      - { $ref: "#/components/parameters/RunId" }
    get:
      tags: [Backtests]
      operationId: getBacktestProgress
      summary: Live progress for one run.
      description: |
        Designed to be polled: about once a second while a run is `RUNNING` or `FINALIZING`, and
        every two and a half seconds otherwise. `sequence` is monotonic per run, so a client can
        drop a response that lost a race, and a terminal run never reopens.

        This is the only polled endpoint in the API and it has a policy of its own sized for that
        cadence. A client that receives `429` here should honour `Retry-After` rather than keep its
        timer — the page it is drawing does not need to blank out, and retrying inside the window
        cannot succeed.
      security: [{ cookieAuth: [] }]
      x-rate-limit: { policy: progress-poll }
      responses:
        "200":
          description: The newest checkpoint.
          headers: *rateLimitHeaders
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BacktestProgress" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/OwnedResourceNotFound" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /backtests/{runId}/strategy:
    parameters:
      - { $ref: "#/components/parameters/RunId" }
    get:
      tags: [Backtests]
      operationId: getBacktestRunStrategy
      summary: The exact strategy definition this run executed.
      description: |
        Read from the run's immutable snapshot, never from the live strategy. It stays correct after
        the strategy has been edited or deleted, which is what makes a result reproducible.
      security: [{ cookieAuth: [] }]
      x-rate-limit: { policy: standard-read }
      responses:
        "200":
          description: The snapshotted strategy name, version number and definition.
          headers: *rateLimitHeaders
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BacktestRunStrategy" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/OwnedResourceNotFound" }
        "429": { $ref: "#/components/responses/TooManyRequests" }

  # ---------------------------------------------------------------------------------------------
  # Benchmarks
  # ---------------------------------------------------------------------------------------------
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: intrinsic_auth
      description: |
        The session cookie issued by `POST /auth/login` or by the Google callback. `HttpOnly`,
        `SameSite=Lax`, and `Secure` in production. Its name is deployment configuration
        (`AUTH_COOKIE_NAME`); `intrinsic_auth` is the default. A browser sends it automatically —
        cross-origin callers must send credentials explicitly, and the deployment's CORS allowlist
        must contain their origin.
    stripeSignature:
      type: apiKey
      in: header
      name: Stripe-Signature
      description: |
        Signature over the raw request body, verified with the webhook signing secret. The only
        authentication the webhook endpoint has, and the only one it needs.
  headers:
    RateLimitPolicy:
      description: The rate-limit policy that produced the numbers in the other headers.
      schema: { type: string }
      example: standard-read
    RateLimitLimit:
      description: The policy's allowance for the current window.
      schema: { type: integer, minimum: 1 }
      example: 240
    RateLimitRemaining:
      description: Allowance left after this request.
      schema: { type: integer, minimum: 0 }
      example: 239
    RateLimitReset:
      description: Whole seconds until the allowance refills.
      schema: { type: integer, minimum: 1 }
      example: 57
    RetryAfter:
      description: |
        Whole seconds to wait before retrying. Always at least 1 — a `Retry-After: 0` would invite
        the immediate retry that caused the refusal. On `429` it is measured from the counter's own
        remaining window, so waiting exactly this long is sufficient.
      schema: { type: integer, minimum: 1 }
      example: 42
  parameters:
...
    StrategyId:
      name: strategyId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    RunId:
      name: runId
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    BadRequest:
      description: The request was malformed or failed validation.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: No session, or a session that is expired, forged or otherwise unusable.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    EntitlementDenied:
      description: |
        The caller's plan does not allow this. The body carries a stable `ENTITLEMENT_*` code plus
        the numbers behind it, so a client can say "your plan allows 10, this change would make 11"
        without parsing the message.

        This is a **product limit**, not a rate limit: retrying identically will fail identically,
        and the remedy is a different plan or a smaller request — never waiting.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/EntitlementError" }
    OwnedResourceNotFound:
      description: |
        No such resource for this caller. A resource that exists but belongs to somebody else
        answers exactly like one that does not exist, so this response reveals nothing about another
        user's data.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    TooManyRequests:
      description: |
        The caller exceeded this endpoint's rate-limit policy. `Retry-After` is the measured time
        until the allowance refills; waiting exactly that long is sufficient, and retrying sooner is
        refused again without consuming anything further.

        A client should surface this as throttling with the wait, never as a generic failure — and
        on `POST /auth/login` in particular it must not be shown as "wrong credentials", because the
        attempt is refused whether or not the password was correct.
      headers:
        Retry-After: { $ref: "#/components/headers/RetryAfter" }
        RateLimit-Policy: { $ref: "#/components/headers/RateLimitPolicy" }
        RateLimit-Limit: { $ref: "#/components/headers/RateLimitLimit" }
        RateLimit-Remaining: { $ref: "#/components/headers/RateLimitRemaining" }
        RateLimit-Reset: { $ref: "#/components/headers/RateLimitReset" }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/RateLimitError" }
          examples:
            search:
              value:
                statusCode: 429
                message: "Too many requests. Retry in 42 seconds."
                code: RATE_LIMITED
                policy: stock-search
                retryAfterSeconds: 42
...
  schemas:
...
    Error:
      type: object
      description: |
        The API's error shape. `code` is present on every refusal a client can reasonably branch on
        and absent on generic framework errors; branch on it, never on `message`.
      required: [statusCode, message]
      properties:
        statusCode: { type: integer, description: Repeats the HTTP status. }
        message:
          {
            type: string,
            description: Human-readable. Never parsed by a client.,
          }
        code:
          {
            type: string,
            description: Stable machine-readable reason,
            when there is one.,
          }
      examples:
        - { statusCode: 404, message: "Stock symbol was not found" }
    RateLimitError:
      allOf:
        - { $ref: "#/components/schemas/Error" }
        - type: object
          required: [code, policy, retryAfterSeconds]
          properties:
            code:
              type: string
              enum: [RATE_LIMITED, RATE_LIMIT_UNAVAILABLE]
              description: |
                `RATE_LIMITED` — the caller exceeded an allowance (`429`).
                `RATE_LIMIT_UNAVAILABLE` — the limiter could not answer and this endpoint's policy
                is fail-closed (`503`). The caller's allowance is untouched.
            policy:
              type: string
              description: |
                Which class of endpoint refused, so a client can tell "your typing is too fast" from
                "you have submitted too many backtests" without matching on the path. Stable
                identifier; the numbers behind it are operational configuration.
              enum:
                - auth-sensitive
                - session-probe
                - stock-search
                - stock-read
                - standard-read
                - progress-poll
                - mutation
                - monitor-mutation
                - backtest-execution
                - billing-mutation
                - admin-operation
            retryAfterSeconds:
              type: integer
              minimum: 1
              description: Mirrors `Retry-After`. Whole seconds, rounded up, never zero.
...
    StrategySignal:
      type: object
      description: |
        Zero or more conditions, ANDed, plus at most one optional trigger which is ANDed with them
        for the same date. A signal with neither is empty and is rejected.
      required: [conditions]
      properties:
        conditions:
          type: array
          maxItems: 10
          items: { $ref: "#/components/schemas/StrategyCondition" }
        trigger: { $ref: "#/components/schemas/StrategyTrigger" }
    StrategyDefinition:
      type: object
      description: |
        The complete strategy. At least one BUY level is required. A level owns exactly one signal
        and one percentage; there is no stock list, no valuation source and no execution parameter
        here — capital, contributions, period and `maximumPositions` all belong to a backtest.
      required: [schemaVersion, buyLevels, sellLevels]
      properties:
        schemaVersion:
          { type: integer, description: Definition schema version. }
        buyLevels:
          type: array
          minItems: 1
          maxItems: 10
          items:
            type: object
            required: [id, signal, percentage]
            properties:
              id: { type: string }
              signal: { $ref: "#/components/schemas/StrategySignal" }
              percentage:
                type: integer
                enum: [25, 50, 75, 100]
                description: Fraction of a full position this level buys.
        sellLevels:
          type: array
          maxItems: 10
          items:
            type: object
            required: [id, signal, percentage]
            properties:
              id: { type: string }
              signal: { $ref: "#/components/schemas/StrategySignal" }
              percentage: { type: integer, enum: [25, 50, 75] }
        finalExit:
          type: object
          description: Closes the whole remaining position when its signal matches.
          required: [id, signal]
          properties:
            id: { type: string }
            signal: { $ref: "#/components/schemas/StrategySignal" }
    StrategySummary:
      type: object
      required:
        [
          id,
          name,
          buyLevelCount,
          sellLevelCount,
          hasFinalExit,
          versionNumber,
          createdAt,
          updatedAt,
        ]
      properties:
        id: { type: string, format: uuid }
        name: { type: string, maxLength: 120 }
        description: { type: string, maxLength: 500 }
        buyLevelCount: { type: integer, minimum: 0 }
        sellLevelCount: { type: integer, minimum: 0 }
        hasFinalExit: { type: boolean }
        versionNumber:
          {
            type: integer,
            minimum: 1,
            description: Incremented by every definition replace.,
          }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    StrategyDetail:
      allOf:
        - { $ref: "#/components/schemas/StrategySummary" }
        - type: object
          required: [definition]
          properties:
            definition: { $ref: "#/components/schemas/StrategyDefinition" }
...
    UpdateStrategyRequest:
      type: object
      minProperties: 1
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        description:
          oneOf:
            - { type: string, maxLength: 500 }
            - { type: "null" }
...
    BacktestRunStatus:
      type: string
      enum: [QUEUED, PREPARING_DATA, RUNNING, FINALIZING, COMPLETED, FAILED]
    CreateBacktestRunRequest:
      type: object
      required:
        [
          strategyId,
          stockListId,
          startDate,
          endDate,
          initialCapital,
          maximumPositions,
        ]
      properties:
        strategyId:
          {
            type: string,
            format: uuid,
            description: Must be owned by the caller.,
          }
        stockListId:
          {
            type: string,
            format: uuid,
            description: Must be owned by the caller and non-empty.,
          }
        benchmarkCode:
          type: string
          description: A catalog benchmark code. Omitting it uses this deployment's default.
          examples: [SP500]
        startDate: { type: string, format: date }
        endDate:
          type: string
          format: date
          description: |
            Must not precede `startDate`. The period may not exceed 30 years, and a plan may bound
            it further (`backtests.maxHistoricalYears`), which is a `403` rather than a `400`.
        initialCapital: { type: number, minimum: 1, maximum: 1000000000 }
        monthlyContribution:
          type: number
          minimum: 0
          maximum: 10000000
          description: Defaults to 0.
        maximumPositions:
          type: integer
          minimum: 1
          maximum: 100
          description: |
            A full position is `1 / maximumPositions` of the portfolio. There is deliberately no
            user-entered maximum-allocation percentage; this is the one input.
...
    BacktestRunConfiguration:
      type: object
      description: |
        The immutable snapshot the run executed. `strategyId` and `stockListId` are nullable because
        the live rows may have been deleted since — the snapshot survives them, which is what makes
        deleting a strategy safe.
      required:
        [
          strategyName,
          strategyVersionNumber,
          stockListName,
          securityCount,
          startDate,
          endDate,
          initialCapital,
          monthlyContribution,
          maximumPositions,
          fullPositionPercent,
          benchmark,
          methodology,
        ]
      properties:
        strategyId:
          oneOf: [{ type: string, format: uuid }, { type: "null" }]
        strategyName: { type: string }
        strategyVersionNumber: { type: integer }
        stockListId:
          oneOf: [{ type: string, format: uuid }, { type: "null" }]
        stockListName: { type: string }
        securityCount: { type: integer, minimum: 0 }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        initialCapital: { type: number }
        monthlyContribution: { type: number }
        maximumPositions: { type: integer }
        fullPositionPercent:
          { type: number, description: "Derived: 100 / maximumPositions." }
        benchmark: { $ref: "#/components/schemas/BacktestBenchmarkSnapshot" }
        methodology: { $ref: "#/components/schemas/BacktestMethodology" }
    NullableNumber:
      oneOf: [{ type: number }, { type: "null" }]
    NullableDate:
      oneOf: [{ type: string, format: date }, { type: "null" }]
    NullableDateTime:
      oneOf: [{ type: string, format: date-time }, { type: "null" }]
    NullableString:
      oneOf: [{ type: string }, { type: "null" }]
    BacktestRunSummary:
      type: object
      required:
        [
          id,
          status,
          strategyName,
          stockListName,
          benchmarkCode,
          benchmarkName,
          startDate,
          endDate,
          initialCapital,
          maximumPositions,
          queuedAt,
          startedAt,
          completedAt,
          progressPercent,
          progressMessage,
          portfolioReturnPercent,
          benchmarkReturnPercent,
          alphaPercent,
        ]
      properties:
        id: { type: string, format: uuid }
        status: { $ref: "#/components/schemas/BacktestRunStatus" }
        strategyName: { type: string }
        stockListName: { type: string }
        benchmarkCode: { type: string }
        benchmarkName: { type: string }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        initialCapital: { type: number }
        maximumPositions: { type: integer }
        queuedAt: { type: string, format: date-time }
        startedAt: { $ref: "#/components/schemas/NullableDateTime" }
        completedAt: { $ref: "#/components/schemas/NullableDateTime" }
        progressPercent: { type: number, minimum: 0, maximum: 100 }
        progressMessage: { $ref: "#/components/schemas/NullableString" }
        portfolioReturnPercent: { $ref: "#/components/schemas/NullableNumber" }
        benchmarkReturnPercent: { $ref: "#/components/schemas/NullableNumber" }
        alphaPercent: { $ref: "#/components/schemas/NullableNumber" }
...
    BacktestMilestone:
      type: object
      description: One completed calendar year of simulation, carrying the prefix computed so far.
      required:
        [
          sequence,
          year,
          simulatedThrough,
          percent,
          completedDays,
          totalDays,
          cash,
          totalValue,
          investedCapital,
          portfolioReturnPercent,
          benchmarkReturnPercent,
          alphaPercent,
          maxDrawdownPercent,
          tradeCount,
          openPositions,
        ]
      properties:
        sequence: { type: integer, minimum: 0 }
        year: { type: string }
        simulatedThrough: { type: string, format: date }
        percent: { type: number, minimum: 0, maximum: 100 }
        completedDays: { type: integer, minimum: 0 }
        totalDays: { type: integer, minimum: 0 }
        cash: { type: number }
        totalValue: { type: number }
        investedCapital: { type: number }
        portfolioReturnPercent: { type: number }
        benchmarkReturnPercent: { $ref: "#/components/schemas/NullableNumber" }
        alphaPercent: { $ref: "#/components/schemas/NullableNumber" }
        maxDrawdownPercent: { type: number }
        tradeCount: { type: integer, minimum: 0 }
        openPositions: { type: integer, minimum: 0 }
...
    BacktestProgress:
      type: object
      required:
        [
          runId,
          status,
          percent,
          message,
          simulatedThrough,
          sequence,
          updatedAt,
          startedAt,
          completedAt,
          live,
          milestones,
          failure,
        ]
      properties:
        runId: { type: string, format: uuid }
        status: { $ref: "#/components/schemas/BacktestRunStatus" }
        percent: { type: number, minimum: 0, maximum: 100 }
        message: { $ref: "#/components/schemas/NullableString" }
        simulatedThrough: { $ref: "#/components/schemas/NullableDate" }
        sequence:
          type: integer
          description: |
            Monotonic per run. A client may safely drop a response whose `sequence` is lower than
            one it has already applied — that is how out-of-order polling responses are handled.
        updatedAt: { $ref: "#/components/schemas/NullableDateTime" }
        startedAt: { $ref: "#/components/schemas/NullableDateTime" }
        completedAt: { $ref: "#/components/schemas/NullableDateTime" }
        live:
          description: Present while a simulation is in flight; null otherwise.
          oneOf:
            - { $ref: "#/components/schemas/BacktestLiveSnapshot" }
            - { type: "null" }
        milestones:
          type: array
          items: { $ref: "#/components/schemas/BacktestMilestone" }
        failure:
          oneOf:
            - { $ref: "#/components/schemas/BacktestFailure" }
            - { type: "null" }
...
    BacktestRunDetail:
      type: object
      required:
        [
          id,
          status,
          configuration,
          queuedAt,
          startedAt,
          completedAt,
          progress,
          live,
          milestones,
          result,
          failure,
        ]
      properties:
        id: { type: string, format: uuid }
        status: { $ref: "#/components/schemas/BacktestRunStatus" }
        configuration: { $ref: "#/components/schemas/BacktestRunConfiguration" }
        queuedAt: { type: string, format: date-time }
        startedAt: { $ref: "#/components/schemas/NullableDateTime" }
        completedAt: { $ref: "#/components/schemas/NullableDateTime" }
        progress:
          type: object
          description: The last persisted checkpoint, so a page reloaded mid-run resumes from it.
          required: [percent, message, simulatedThrough, sequence, updatedAt]
          properties:
            percent: { type: number, minimum: 0, maximum: 100 }
            message: { $ref: "#/components/schemas/NullableString" }
            simulatedThrough: { $ref: "#/components/schemas/NullableDate" }
            sequence: { type: integer }
            updatedAt: { $ref: "#/components/schemas/NullableDateTime" }
        live:
          oneOf:
            - { $ref: "#/components/schemas/BacktestLiveSnapshot" }
            - { type: "null" }
        milestones:
          type: array
          items: { $ref: "#/components/schemas/BacktestMilestone" }
        result:
          description: Present only for `COMPLETED`.
          oneOf:
            - { $ref: "#/components/schemas/BacktestResult" }
            - { type: "null" }
        failure:
          description: Present only for `FAILED`.
          oneOf:
            - { $ref: "#/components/schemas/BacktestFailure" }
            - { type: "null" }
    BacktestRunStrategy:
      type: object
      required: [strategyName, versionNumber, definition]
      properties:
        strategyName: { type: string }
        versionNumber: { type: integer }
        definition: { $ref: "#/components/schemas/StrategyDefinition" }

    # -------------------------------------------------------------------------------------------
    # Monitors
    # -------------------------------------------------------------------------------------------
...